flask/flaskrest/requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
flask-restx==0.3.0
Flask-SQLAlchemy==2.5.1
bcrypt==3.2.0
PyJWT==1.7.1
uWSGI==2.0.19.1
mysqlclient==2.0.3